refactor: 상품 정보 조회 함수 통일화
