+++ backend/src/product/product.repository.ts
@@ -2,11 +2,11 @@ import { Injectable } from '@nestjs/common';
 import { Repository } from 'typeorm';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Product } from 'src/entities/product.entity';
-import { createUrl11st } from 'src/utils/openapi.11st';
 import { ProductDto } from 'src/dto/product.dto';
 import { ProductInfoDto } from 'src/dto/product.info.dto';
 import { MAX_TRACKING_RANK } from 'src/constants';
 import { TrackingProduct } from 'src/entities/trackingProduct.entity';
+import { createUrl } from 'src/utils/product.info';
 
 @Injectable()
 export class ProductRepository extends Repository<Product> {
@@ -19,7 +19,7 @@ export class ProductRepository extends Repository<Product> {
 
     async saveProduct(productDto: ProductDto | ProductInfoDto): Promise<Product> {
         const { productName, productCode, shop, imageUrl } = productDto;
-        const shopUrl = createUrl11st(productCode);
+        const shopUrl = createUrl(shop, productCode);
         const newProduct = Product.create({ productName, productCode, shop, shopUrl, imageUrl });
         await newProduct.save();
         return newProduct;
